Use u16 for memory addresses

+++ crates/cpu/src/instructions.rs
@@ -63,11 +63,11 @@ impl Condition {
 }
 
 fn get_8bit_operand(pc: u16, memory: &Memory) -> u8 {
-    memory.get((pc as usize) + 1)
+    memory.get(pc + 1)
 }
 
 fn get_16bit_operand(pc: u16, memory: &Memory) -> u16 {
-    memory.get_u16((pc as usize) + 1)
+    memory.get_u16(pc + 1)
 }
 
 /// Decode an operation code into an `Instruction`.
@@ -272,7 +272,7 @@ mod tests {
             (0b0011_0111, I::RotateShift(RS::SWAP(0b0011_0111))),
             (0b0011_1111, I::RotateShift(RS::SRL(0b0011_1111))),
         ] {
-            memory.set((pc as usize) + 1, operand);
+            memory.set(pc + 1, operand);
 
             assert_eq!(
                 instruction,
@@ -293,7 +293,7 @@ mod tests {
             (0b1100_1111, I::Bit(Bit::SET(0b1100_1111))),
             (0b1001_0111, I::Bit(Bit::RES(0b1001_0111))),
         ] {
-            memory.set((pc as usize) + 1, operand);
+            memory.set(pc + 1, operand);
 
             assert_eq!(
                 instruction,
